Fix build error

+++ src/server/shared/Networking/Socket.h
@@ -256,7 +256,7 @@ private:
 
         const uint8 addressFamily = readPointer[13];
         const uint16 len = (readPointer[14] << 8) | readPointer[15];
-        if (len+16 > packet.GetActiveSize())
+        if (static_cast<size_t>(len) + 16 > packet.GetActiveSize())
         {
             AsyncReadProxyHeader();
             return;
